Hooked up api caller
Have a model helper that now calls the api and returns a card or cards,
successfully triggering the mana cost translator.

pulled the mana cost translation logic out of the manaCost model and
into a translator class with extension methods.

updated unit tests to accommodate the changes and to account for a
simple integration test.